move components to new file
